pyproject.toml
[project]
name = "mcp-client-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.7.1",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
]
